v14. AOP. Iteration 1.

+++ hw05-AOPprincipes/src/main/java/aop/Useful.java
@@ -3,6 +3,6 @@ package aop;
 public interface Useful {
     public void sayHelloTo();
     public void sayHelloTo(String name);
-    public void sayHelloTo(String name, short howMany);
+    public void sayHelloTo(String name, double howMany);
 }
